combine FieldParameters and Field in a signle class, cleanup
