+++ field_friend/automations/__init__.py
@@ -2,8 +2,7 @@
 from .automation_watcher import AutomationWatcher
 from .battery_watcher import BatteryWatcher
 from .coverage_planer import CoveragePlanner
-from .field import Field, FieldObstacle, Row
-from .field_parameters import FieldParameters
+from .field import Field, Row
 from .field_provider import FieldProvider
 from .implements.implement import Implement
 from .kpi_provider import KpiProvider
@@ -20,8 +19,6 @@ __all__ = [
     'CoveragePlanner',
     'DetectorError',
     'Field',
-    'FieldObstacle',
-    'Field_Parameters',
     'FieldProvider',
     'Implement',
     'CoverageNavigation',
