+++ field_friend/interface/pages/__init__.py
@@ -1,5 +1,4 @@
 from .dev_page import dev_page
-from .field_planner_page import field_planner_page
 from .kpi_page import kpi_page
 from .main_page import main_page
 from .monitor_page import monitor_page
